Thermistor temperature array
Calculated and added thermistor temperature values in a header. As the equation of temperature-resistance for the thermistor is roughly quartic, the adc is only 8 bits, and we have a large amount of memory available, it makes sense to just calculate and store the temperature for each reading of the adc. As the adc is not particularly accurate, all temperatures are stored rounded to the nearest degree. Temperature in degrees celcius.